+++ tests/conftest.py
@@ -17,3 +17,14 @@ def cfradial1_file(tmp_path_factory):
     fname = os.path.join(fn, "cfradial1_data.nc")
     urlretrieve(url, filename=fname)
     return fname
+
+
+@pytest.fixture(scope="session")
+def odim_file(tmp_path_factory):
+    base_url = "https://raw.githubusercontent.com/wradlib/wradlib-data/main/hdf5/"
+    filename = "71_20181220_060628.pvol.h5"
+    url = urljoin(base_url, filename)
+    fn = tmp_path_factory.mktemp("odim_data")
+    fname = os.path.join(fn, "odim_data.h5")
+    urlretrieve(url, filename=fname)
+    return fname