+++ xradar/io/backends/__init__.py
@@ -14,5 +14,6 @@ Xarray Backends
 """
 
 from .cfradial1 import *  # noqa
+from .odim import *  # noqa
 
 __all__ = [s for s in dir() if not s.startswith("_")]